Rearranged the lambda structure
